requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
python-dotenv==1.0.1
openpyxl==3.1.5
